+++ src/uk/ac/cam/eng/rulebuilding/features/RuleCount1.java
@@ -23,7 +23,7 @@ public class RuleCount1 implements Feature {
     @Override
     public double value(Rule r, ArrayWritable mapReduceFeatures) {
         double count = ((DoubleWritable) mapReduceFeatures.get()[2]).get();
-        return (count == 1) ? 1 : 0;
+        return (int) ((count == 1) ? 1 : 0);
     }
 
 }
